perf: restructuring error dashboard components to eliminate duplicate api calls
