--- src/errors/pages/ErrorDashboardLayout.jsx
@@ -1,27 +1,0 @@
-import React from "react";
-import { Outlet } from "react-router-dom";
-import { Grid, GridItem } from "@chakra-ui/react";
-
-import Sidebar from "../components/Sidebar";
-
-export default function ErrorDashboardLayout(props) {
-  return (
-    <div className="site-wrapper">
-      <Grid templateColumns="repeat(6, 1fr)" bg="gray.50">
-        <GridItem
-          as="aside"
-          colSpan={{ base: 6, lg: 2, xl: 1 }}
-          bg="purple.400"
-          minHeight={{ lg: "100vh" }}
-          p={{ base: "20px", lg: "30px" }}
-        >
-          <Sidebar {...props} />
-        </GridItem>
-
-        <GridItem as="main" colSpan={{ base: 6, lg: 4, xl: 5 }} p="40px">
-          <Outlet />
-        </GridItem>
-      </Grid>
-    </div>
-  );
-}
